refactor: 공고상세페이지(로그인 상태): 컴포넌트 분리
